@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Optional, Self, Union, Callable
+from typing import Optional, Self, Callable
 from numpy import typing as npt
 
 
@@ -12,17 +12,17 @@ class Scale:
 class Normalize(Scale):
     """Normalize the data so that it fits in a bounding box defined by bbox_min and bbox_max."""
 
-    bbox_min: Union[npt.NDArray, float]
-    bbox_max: Union[npt.NDArray, float]
-    value_min: Union[npt.NDArray, float]
-    value_max: Union[npt.NDArray, float]
+    bbox_min: npt.NDArray | float
+    bbox_max: npt.NDArray | float
+    _value_min: npt.NDArray | float | None = None
+    _value_max: npt.NDArray | float | None = None
 
 
 @dataclass(kw_only=True)
 class Log(Scale):
     """Logarithmic scale."""
 
-    base: float
+    base: float = 10.0
 
 
 @dataclass(kw_only=True)
